requirements.txt
# Core dependencies - Compatible versions
python-telegram-bot[webhooks]==21.0.1
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0
asyncpg==0.29.0

# AI and Image Processing
openai==1.3.8
Pillow==10.1.0
python-multipart==0.0.6

# Payment Processing (Tether only for now)
web3==6.11.3

# Utilities - Compatible versions
python-dotenv==1.0.0
httpx>=0.27.0
redis==5.0.1
aiofiles==23.2.1
pydantic==2.5.0
pydantic-settings==2.1.0

# Development (minimal)
pytest==7.4.3
pytest-asyncio==0.21.1
